+++ AEther/DFT/WindowedDFTFilterSIMD.cs
@@ -14,39 +14,39 @@ namespace AEther
 
         public int Length => Buffer.Size;
 
-        ComplexVector<float> States;
-        readonly ComplexVector<float> Coefficients;
-        readonly float Frequency;
-        readonly float DFTFrequency;
-        readonly float Q;
+        ComplexVector<double> States;
+        readonly ComplexVector<double> Coefficients;
+        readonly double Frequency;
+        readonly double DFTFrequency;
+        readonly double Q;
         readonly int M;
-        readonly Vector<float> Window;
+        readonly Vector<double> Window;
         readonly int WindowSize;
-        readonly RingBuffer<float> Buffer;
+        readonly RingBuffer<double> Buffer;
 
-        public WindowedDFTFilterSIMD(float frequency, float frequencyResolution, float sampleRate, float[] window)
+        public WindowedDFTFilterSIMD(double frequency, double frequencyResolution, double sampleRate, double[] window)
         {
 
             WindowSize = window.Length;
-            var windowPadded = window.Concat(Enumerable.Repeat(0f, Vector<float>.Count - window.Length)).ToArray();
-            Window = new Vector<float>(windowPadded);
+            var windowPadded = window.Concat(Enumerable.Repeat(0.0, Vector<double>.Count - window.Length)).ToArray();
+            Window = new Vector<double>(windowPadded);
             Frequency = frequency;
             DFTFrequency = Frequency / sampleRate;
-            Q = 1f / (float)(Math.Pow(2, 1f / frequencyResolution) - 1);
+            Q = 1.0 / (Math.Pow(2, 1.0 / frequencyResolution) - 1);
             M = (int)Math.Round(Q);
-            States = ComplexVector<float>.Zero;
+            States = ComplexVector<double>.Zero;
             var length = (int)Math.Round(M / DFTFrequency);
             Buffer = new(length);
-            Coefficients = ComplexVector<float>.Zero;
-            var coeff = new Complex[ComplexVector<float>.Count];
+            Coefficients = ComplexVector<double>.Zero;
+            var coeff = new Complex[ComplexVector<double>.Count];
             for (int j = 0; j < window.Length; ++j)
             {
                 var bin = M - (window.Length / 2) + 1 + j;
                 coeff[j] = Complex.Exp(2 * Math.PI * Complex.ImaginaryOne * bin / Length);
             }
-            var coeffRe = new Vector<float>(coeff.Select(c => (float)c.Real).ToArray());
-            var coeffIm = new Vector<float>(coeff.Select(c => (float)c.Imaginary).ToArray());
-            Coefficients = new ComplexVector<float>(coeffRe, coeffIm);
+            var coeffRe = new Vector<double>(coeff.Select(c => c.Real).ToArray());
+            var coeffIm = new Vector<double>(coeff.Select(c => c.Imaginary).ToArray());
+            Coefficients = new ComplexVector<double>(coeffRe, coeffIm);
 
         }
 
@@ -62,21 +62,21 @@ namespace AEther
             return result / Buffer.Size;
         }
 
-        public void Process(float newSample)
+        public void Process(double newSample)
         {
             var oldSample = Buffer.Add(newSample);
-            var input = new Vector<float>(newSample - oldSample);
+            var input = new Vector<double>(newSample - oldSample);
             States = Coefficients * States + input;
         }
 
-        public void Process(ReadOnlyMemory<float> samples)
+        public void Process(ReadOnlyMemory<double> samples)
         {
             while (0 < samples.Length)
             {
                 var dst = Buffer.GetMemory(samples.Length);
                 for (var i = 0; i < dst.Length; ++i)
                 {
-                    var input = new Vector<float>(samples.Span[i] - dst.Span[i]);
+                    var input = new Vector<double>(samples.Span[i] - dst.Span[i]);
                     States = Coefficients * States + input;
                 }
                 samples[0..dst.Length].CopyTo(dst);
