+++ AEther/Domain.cs
@@ -9,27 +9,27 @@ using System.Xml.Serialization;
 
 namespace AEther
 {
-    public class Domain : IEnumerable<float>
+    public class Domain : IEnumerable<double>
     {
 
         public int Count => Values.Length;
         public int Resolution;
-        public float Octaves => Count / Resolution;
-        public float MinValue => Values[0];
+        public double Octaves => Count / Resolution;
+        public double MinValue => Values[0];
 
-        readonly float[] Values;
+        readonly double[] Values;
 
-        public float this[int i] => Values[i];
+        public double this[int i] => Values[i];
 
-        public Domain(float minValue, int resolution, int count)
+        public Domain(double minValue, int resolution, int count)
         {
             Values = Enumerable.Range(0, count)
-                .Select(i => minValue * (float)Math.Pow(2, i / (double)resolution))
+                .Select(i => minValue * Math.Pow(2, i / (double)resolution))
                 .ToArray();
             Resolution = resolution;
         }
 
-        public static Domain FromRange(float minValue, float maxValue, int resolution)
+        public static Domain FromRange(double minValue, double maxValue, int resolution)
         {
             int count = (int)Math.Round(resolution * Math.Log(maxValue / minValue, 2));
             return new Domain(minValue, resolution, count);
@@ -52,11 +52,11 @@ namespace AEther
                 ^ (234567891 * Count.GetHashCode())
                 ^ (345678912 * MinValue.GetHashCode());
 
-        public IEnumerator<float> GetEnumerator()
-            => ((IEnumerable<float>)Values).GetEnumerator();
+        public IEnumerator<double> GetEnumerator()
+            => ((IEnumerable<double>)Values).GetEnumerator();
 
         IEnumerator IEnumerable.GetEnumerator()
-            => ((IEnumerable<float>)Values).GetEnumerator();
+            => ((IEnumerable<double>)Values).GetEnumerator();
 
     }
 }
